requirements.txt
Flask
Flask-SQLAlchemy
requests
python-dotenv
wasmtime
pycryptodome
